Create Convertor
Intial Commit

Just made a simple GUI that sort of works. Needs muhc imporvmenets. Accepts pdf files and displays the file path. Convertor button works but i need to create a fucntion so that when its pressed to works. I also want the file path to display in the box itself.