Update Convertor
Changed much of the gui in order to split it up into 6ths. More chages needed as scaling of text is wonky. I changed it so that the boxs scale depenidng on the users windwo size. More changes to GUI hopefully. Akash needs to read doumentation but its going well