Added conversion method to code
This is a huge upate. Realisticly i created a class which is meant to convert the pdf into a tns file.  Comments were als added to every line in order to amke it clear to the user what is going. I also made a loop that goes thorugh each element in tht list allowing for all the files to be converted.